Added threading to client to support 2 blocking IO operations

+++ universal/packet_format.h
@@ -29,7 +29,7 @@ struct packet{
 };
 
 struct encoded_packet{
-    uint16_t length;
+    uint32_t length;
     char *text;
 };
